data_structures/charts/python-ms/requirements.txt
Flask==2.0.1
Flask-SocketIO==5.2.0
Flask-Cors==4.0.2
requests==2.25.1
python-binance==1.0.15
werkzeug==2.0.3